+++ packages/wasm/tests/types/types-node.test.ts
@@ -1,7 +1,9 @@
-import { createMuhammaraWasm } from "../../index.js";
+import { createMuhammaraWasm, createRecipe } from "../../index.js";
 
 async function usesNodeCompatibleTypes() {
   var muhammara = await createMuhammaraWasm();
+  await createRecipe({ defaultFont: Buffer.alloc(0) });
+  await createRecipe({ defaultFont: false });
   var writer = muhammara.createWriter();
   var output = new muhammara.PDFWStreamForBuffer();